refactor: libs.Automerge
