+++ code/sys.driver/ext.lib.automerge/src/common/libs.ts
@@ -1,8 +1,8 @@
 /**
  * @external
  */
-export { next as A } from '@automerge/automerge';
 export { PatchState } from 'ext.lib.immer';
+export { next as A } from '@automerge/automerge';
 
 import { clone, equals, uniq, uniqBy } from 'ramda';
 export const R = { clone, equals, uniq, uniqBy } as const;
